updated Modulo and Remainder; updated comments

+++ Runtime/Scripts/System/Extensions/Numerics/Integrals/Byte/ByteExtensions.Modulo.cs
@@ -6,9 +6,29 @@ namespace WellDefinedValues
 
 	public static partial class ByteExtensions
 	{
+		/// <remarks>
+		/// Repeats the sequence from <c>zero</c> to <c>divisor</c> (exclusively), 
+		/// with the direction and <c>sign</c> based on the <c>sign</c> of the <c>divisor</c>.
+		/// </remarks>
+		/// <seealso cref="Remainder"/>
+		/// <example>
+		/// <code>
+		/// int modulo = 3;
+		/// int[] input    = { -5, -4, -3, -2, -1, 0, 1, 2, 3, 4, 5 };
+		/// int[] expected = {  1,  2,  0,  1,  2, 0, 1, 2, 0, 1, 2 };
+		/// int[] actual   = {  1,  2,  0,  1,  2, 0, 1, 2, 0, 1, 2 };
+		/// </code>
+		/// 
+		/// <code>
+		/// int modulo = -3;
+		/// int[] input    = { -5, -4, -3, -2, -1, 0,  1,  2, 3,  4,  5 };
+		/// int[] expected = { -2, -1,  0, -2, -1, 0, -2, -1, 0, -2, -1 };
+		/// int[] actual   = { -2, -1,  0, -2, -1, 0, -2, -1, 0, -2, -1 };
+		/// </code>
+		/// </example>
 		public static byte Modulo(this byte dividend, byte divisor)
 		{
-			if(divisor == Int.Zero)
+			if(divisor == Byte.Zero)
 			{
 				throw new DivideByZeroException(nameof(Modulo) + "(0) is undefined.");
 			}
